disable musical-specific thing

+++ src/lib/score/model/tree/TreeNode.hpp
@@ -14,7 +14,7 @@
 template <typename T>
 auto& child_at(const std::list<T>& list, int index)
 {
-  SCORE_ASSERT(index >= 0 && index < list.size());
+  SCORE_ASSERT(index >= 0 && index < (int)list.size());
   auto it = list.begin();
   std::advance(it, index);
   return *it;
@@ -23,7 +23,7 @@ auto& child_at(const std::list<T>& list, int index)
 template <typename T>
 auto& child_at(std::list<T>& list, int index)
 {
-  SCORE_ASSERT(index >= 0 && index < list.size());
+  SCORE_ASSERT(index >= 0 && index < (int)list.size());
   auto it = list.begin();
   std::advance(it, index);
   return *it;
